requirements.txt
asn1crypto>=0.24.0  
certifi>=2019.3.9
cffi>=1.12.2  
chardet>=3.0.4   
cryptography>=2.6.1   
enum34>=1.1.6   
idna>=2.8     
pip>=19.0.3  
pycparser>=2.19    
PyKMIP>=0.8.0   
requests>=2.21.0  
setuptools>=40.8.0  
six>=1.12.0  
SQLAlchemy>=1.3.1   
urllib3>=1.24.1  
wheel>=0.33.1  
